\name{readSvmLight}
\alias{readSvmLight}
\title{Reads a sparse matrix from a SVMlight compatible file}
\usage{
  readSvmLight(filename)
}
\arguments{
  \item{fileName}{input file name}
}
\value{
  list with a sparse matrix and a list of labels
}
\description{
  Reads a sparse matrix from a SVMlight compatible file
}

