\name{write.svmlight}
\alias{write.svmlight}
\title{Write a sparse matrix to a file in  SVMlight format}
\usage{
  write.svmlight(sparseMatrix, labelVector, file)
}
\arguments{
  \item{sparseMatrix}{}

  \item{labelVector}{}

  \item{file}{}
}
\description{
  Write a sparse matrix to a file in SVMlight format
}

