\name{writeSvmLight}
\alias{writeSvmLight}
\title{Writes a sparse matrix to a SVMlight compatible file}
\usage{
  writeSvmLight(inputMatrix, labels, fileName)
}
\arguments{
  \item{inputMatrix}{sparse matrix}

  \item{labels}{list of numeric labels for each row in the
  matrix}

  \item{fileName}{output file name}
}
\value{
  list with debug information
}
\description{
  Writes a sparse matrix to a SVMlight compatible file
}

